+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [datafactory] - Databricks clusters and instance pools can be configured to be added as linked services
 - [dbt] - Added containers for `models` and `snapshots`
 - [devcontainer] - Additional extensions to improve the dev experience
+- [devops] - Better defaults when defining the virtual machine scale set
 - [docker_image/iac_runtime] - Create a new Dockerfile. Use the same IaC runtime for the platform deployment and for the devcontainers.
 - [platform/cicd] - Making use of the built in source code in the Docker image (/platform/src).
 - [platform/source] - Publish the source of the platform in the Docker image (/platform/src). Make the Docker image `private` in the dockerhub.